rename: 패키지 이름 response -> result로 수정
