+++ app/src/main/java/woowacourse/shopping/domain/repository/CouponRepository.kt
@@ -1,7 +1,7 @@
 package woowacourse.shopping.domain.repository
 
 import woowacourse.shopping.domain.model.coupon.Coupon
-import woowacourse.shopping.domain.response.Response
+import woowacourse.shopping.domain.result.Response
 
 interface CouponRepository {
     suspend fun allCoupons(): List<Coupon>
