+++ app/src/main/java/woowacourse/shopping/ui/cart/cartitem/uimodel/CartError.kt
@@ -1,11 +1,5 @@
 package woowacourse.shopping.ui.cart.cartitem.uimodel
 
-import android.util.Log
-import woowacourse.shopping.domain.model.CartWithProduct
-import woowacourse.shopping.domain.model.Product
-import woowacourse.shopping.domain.response.Fail
-import woowacourse.shopping.domain.response.Response
-
 sealed interface CartError {
     data object UpdateCart : CartError
