+++ app/src/main/java/woowacourse/shopping/ui/coupon/uimodel/CouponError.kt
@@ -1,14 +1,5 @@
 package woowacourse.shopping.ui.coupon.uimodel
 
-import android.util.Log
-import woowacourse.shopping.domain.model.CartWithProduct
-import woowacourse.shopping.domain.model.Order
-import woowacourse.shopping.domain.model.Product
-import woowacourse.shopping.domain.model.coupon.Coupon
-import woowacourse.shopping.domain.response.Fail
-import woowacourse.shopping.domain.response.Response
-import woowacourse.shopping.ui.cart.cartitem.uimodel.CartError
-
 sealed interface CouponError {
     data object Order : CouponError
