+++ app/src/main/java/woowacourse/shopping/ui/detail/uimodel/ProductDetailError.kt
@@ -1,13 +1,5 @@
 package woowacourse.shopping.ui.detail.uimodel
 
-import android.util.Log
-import woowacourse.shopping.data.local.room.recentproduct.RecentProduct
-import woowacourse.shopping.domain.model.CartWithProduct
-import woowacourse.shopping.domain.model.Product
-import woowacourse.shopping.domain.response.Fail
-import woowacourse.shopping.domain.response.Response
-import woowacourse.shopping.ui.cart.cartitem.uimodel.CartError
-
 sealed interface ProductDetailError {
     data object LoadProduct:ProductDetailError
